test: cover installer status wrapper
